fix: 패키지 변경, 파일 이름 변경, 일부 given, when, then 적용
